task table upgrade
The task table is upgraded on client side with tansack's table.
On server side, api were renamed and refactored in taskcontroller.
Now users can select multiple tasks and change their status with one click.
Also multiple fields of description and/or status can be edited manually and updated to database with one click.
The table is sorting enabled now.